+++ src/main/java/com/kraskovskij/car_shop/service/implementations/CarServiceImpl.java
@@ -29,7 +29,7 @@ public class CarServiceImpl implements CarService{
     @Transactional(readOnly = true)
     @Override
     public List<Car> getCarsByParams(String mark, String type, String engineType, int stYear, int endYear,
-                                     double stPrice, double endPrice, int stMileage, int endMileage){
+                                     int stPrice, int endPrice, int stMileage, int endMileage){
         return carRepository.getCarsByParams(mark, type, engineType, stYear, endYear, stPrice, endPrice, stMileage, endMileage);
     }
