+++ src/main/java/com/kraskovskij/car_shop/service/interfaces/CarService.java
@@ -9,7 +9,7 @@ public interface CarService {
 
     List<Car> getAllCars();
 
-    List<Car> getCarsByParams(String mark, String type, String engineType, int stYear, int endYear, double stPrice, double endPrice, int stMileage, int endMileage);
+    List<Car> getCarsByParams(String mark, String type, String engineType, int stYear, int endYear, int stPrice, int endPrice, int stMileage, int endMileage);
 
     void saveCar(Car car);
